Fix type error when running on pipeline

+++ scheduler/include/Scheduler.h
@@ -1,6 +1,7 @@
 #ifndef SCHEDULER_H
 #define SCHEDULER_H
 
+#include <cstdint>
 #include <vector>
 
 class Scheduler
@@ -11,13 +12,13 @@ public:
 
     struct Course
     {
-        size_t Id;
+        uint32_t Id;
         std::vector<Section> Sections;
     };
 
     struct Section
     {
-        size_t Id;
+        uint32_t Id;
         Course* Parent;
         std::vector<Timeslot> Timeslots;
 
@@ -35,7 +36,7 @@ public:
             FRI = 0x10
         };
 
-        size_t Id;
+        uint32_t Id;
         Section* Parent;
         uint16_t StartTime = 0;
         uint16_t EndTime = 0;
